// src/game.services/profile.services/profile.interface.ts
export interface StatPoints {
    mainVocalist: number;
    rapper: number;
    leadDancer: number;
    leadVocalist: number;
    mainDancer: number;
    visual: number;
    leadRapper: number;
}

/**
 * Represents the result when a stats update fails.
 *
 * @interface UpdateStatsFailed
 * @property {boolean} success - Indicates whether the update failed.
 * @property {string} message - A message providing details about the failure.
 */
export interface UpdateStatsFailed {
    success: boolean;
    message: string;
  }
// ...
export interface ProfilePicture {
    userName: string;
    fileFormat: string;
    profilePicture: string;
    uploadedAt: number;
    fileSize: number;
    likes: PictureLikes[];
    id?: string
}


export interface PictureLikes {
    userName?: string;
    timestamp?: number;
    likeId: string; 
}
// ...
export interface SoulMetaData {
	genre1: string;
	genre2: string;
	genre3: string;
	animal1: string;
    animal2: string;
    animal3: string;
	horoscope: string;
    id?: string
    ownership?: string[];
    horoscopeMatch?: string[];
    animalMatch?: string[];
    weeklyFirst?: string[];
}
// ...
export interface CardCollection {
	card: CardMetaData;
	name: string;
	count: number;
}


/**
 * Represents buffer data in string format.
 *
 * @interface BufferData
 * @property {string} bufferData - The raw data in string format.
 */
export interface BufferData {
    bufferData: string;
  }
// ...
  export interface MyNote {
    userName?: string;
    note: string;
    createdAt?: number;
    updatedAt?: number;
  }