// src/user.services/user.service.interface.ts
export interface PlayerStats {
    level: number;
    playerExp: number;
    availStatPoints: number;
    rank: string;
    statPointsSaved: {
      mainVocalist: number;
      rapper: number;
      leadDancer: number;
      leadVocalist: number;
      mainDancer: number;
      visual: number;
      leadRapper: number;
    };
  }