// src/game.services/inventory.services/inventory.interface.ts
export interface CardMetaData {
    name: string;
    description: string;
    group: string;
    era: string;
    scoreBoost: string;
    healBoost: string;
    skill: string;
    stars: string;
    slot: string;
    tier: string;
    boostCount: string;
    awakenCount: string;
    breakthrough: boolean;
    skillEquipped: boolean;
    artist: string; 
    position: string;
    position2: string;
    rarity: string;
    level: string;
    zodiac: string;
    animal: string;
    experience: string;
    currentExperience: string;
    image: string;
    imageByte?: string;

    uri: string;
    owner: string;
    tokenId: string;
    contractAddress: string;
    id: string;
    listingId: number;
    cardAddress: string;
    uploader: string;
    supply: string;
    quantity: 1
    quantityOwned: string
    lister?: string;
    pricePerToken?: number
    currencyName?: string
    transferred?: string
 };